Code refactor Added more documentation.
