+++ src/Contexts/TokenWorkItemContext.ts
@@ -4,6 +4,9 @@ import { Token } from '../Classes/Token';
 import { FlowElement } from '../Classes/FlowElement';
 import { WorkItem } from '../Classes/WorkItem';
 
+/**
+ * Contextual information for customizations of business processes such as changes in form fields.
+ */
 export class TokenWorkItemContext extends Context {
     private token: Token;
     private workitem: WorkItem;
@@ -26,6 +29,9 @@ export class TokenWorkItemContext extends Context {
     }
 }
 
+/**
+ * Contextual information for customizations Sequence flows.
+ */
 export class TokenWorkItemExecuteContext extends TokenWorkItemContext {
     private source: FlowElement;
     private target: FlowElement;
@@ -45,6 +51,9 @@ export class TokenWorkItemExecuteContext extends TokenWorkItemContext {
     }
 }
 
+/**
+ * Contextual information for customizations start events generated by the process engine such as Conditional and Web service events. It can be used to abort the creation of a business process instance.
+ */
 export class TokenWorkItemBeforeInitContext extends Context {
 
     private payload: any;
